Separate verifying event listener requirements
Makes it easier to test and declutters the registry service
